refactor: __init__.py for service
