+++ service/sections.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from sqlalchemy.orm import Session
-from service.db_engine import db_engine
+from service import db_engine
 from src.model import Sections, SectionOperators
